Removing controller-related code in order to remove unnecessary code
